correctionError

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
         BmiService service = new BmiService();
-        double weight = 84.0;
+        int weight = 84;
         double height = 1.78;
         double bmi = service.calculate(weight, height);
         System.out.println("Ваш вес на сегодня:" + weight + " кг");
